requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
branca==0.7.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
executing==2.0.1
fastapi==0.104.1
folium==0.15.0
h11==0.14.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.8
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandarallel==1.6.5
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pydantic==2.5.0
pydantic_core==2.14.1
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-toolbelt==1.0.0
rubicon-objc==0.4.7
selenium==4.15.2
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
streamsync==0.2.8
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.24.0.post1
watchdog==2.3.1
wcwidth==0.2.10
websockets==10.4
widgetsnbextension==4.0.9
wsproto==1.2.0
